[FT11] - Remodelagem do modelo de dados
User Stories: nenhum
Tarefas: Bloquear tentativa de exclusão do admin default.
Bugs: Corrigidas as opções de conexão do Mongo na classe ConnectionFactoryMongodb
Outras informações: Alterado o tipo dos atributos dos elementos do modelo que usavam ENUM, foi alterado para string, e a camada de validação garante a execução correta.
Adicionado um novo atributo 'protected' no modelo User, booleano utilizado para bloquear a exclusão do Admin registrado na primeira execução.
Foram refatorados alguns testes referentes a mudança dos atributos do tipo ENUM e criado um novo para a nova feature

+++ src/application/service/user.service.ts
@@ -14,6 +14,7 @@ import { IntegrationEvent } from '../integration-event/event/integration.event'
 import { IIntegrationEventRepository } from '../port/integration.event.repository.interface'
 import { ILogger } from '../../utils/custom.logger'
 import { UserDeleteEvent } from '../integration-event/event/user.delete.event'
+import { ValidationException } from '../domain/exception/validation.exception'
 
 /**
  * Implementing User Service.
@@ -42,6 +43,13 @@ export class UserService implements IUserService {
         try {
             ObjectIdValidator.validate(id)
             const user: User = await this._userRepository.findOne(new Query().fromJSON({ filters: { _id: id } }))
+
+            if (user && user.type === UserType.ADMIN && user.protected) {
+                throw new ValidationException(
+                    'The operation could not be completed as the user in question cannot be removed.'
+                )
+            }
+
             const result: boolean = await this._userRepository.delete(id)
             if (result) {
                 if (user && user.type !== UserType.ADMIN) {
